move remaining functions from utils into separate submodules
