+++ corebehrt/functional/preparation/utils.py
@@ -1,3 +1,4 @@
+from dataclasses import asdict, fields
 from datetime import datetime
 from typing import List, Union
 
@@ -45,3 +46,22 @@ def get_non_priority_tokens(vocabulary: dict, low_priority_prefixes: List[str]) 
         for k, v in vocabulary.items()
         if any(k.startswith(prefix) for prefix in low_priority_prefixes)
     }
+
+
+def subset_patient_data(patient: PatientData, keep_indices: List[int]) -> PatientData:
+    """
+    Return a new PatientData containing only the rows at `keep_indices`
+    for all list-type attributes. Non-list attributes remain unchanged.
+    """
+    # Convert the PatientData instance to a dictionary
+    data = asdict(patient)
+
+    # For each field in the dataclass, if the value is a list, subset it.
+    # Otherwise, keep it as is.
+    for f in fields(PatientData):
+        val = data[f.name]
+        if isinstance(val, list):
+            data[f.name] = [val[i] for i in keep_indices]
+
+    # Recreate a new PatientData from the updated dictionary
+    return PatientData(**data)
