+++ corebehrt/modules/cohort_handling/index_dates.py
@@ -6,7 +6,7 @@ import numpy as np
 import pandas as pd
 
 from corebehrt.functional.constants import PID_COL, TIMESTAMP_COL
-from corebehrt.functional.utils import filter_table_by_pids
+from corebehrt.functional.preparation.filter import filter_table_by_pids
 
 logger = logging.getLogger(__name__)
